MAINT: Moved `publish` function to publication app; publication app is no longer imported in manager
