+++ splinter_tests/test_publishing.py
@@ -216,7 +216,7 @@ def test_see_published_by_others(user_alice_logged_in, user_bob, handle_usage_st
     surface_description = "Valuable results."
     surface = SurfaceFactory(creator=user_bob, name=surface_name, description=surface_description)
     Topography2DFactory(surface=surface)
-    publication = surface.publish('cc0-1.0', 'Bob')
+    publication = Publication.publish(surface, 'cc0-1.0', 'Bob')
 
     # Alice filters for published surfaces - enters
     # "Select" tab and chooses "Only published surfaces"
@@ -253,7 +253,7 @@ def test_see_published_by_others(user_alice_logged_in, user_bob, handle_usage_st
     #
     # Bob publishes again, Alice reloads the page
     #
-    publication = surface.publish('cc0-1.0', 'Bob')
+    publication = Publication.publish(surface, 'cc0-1.0', 'Bob')
     browser.reload()
 
     #
@@ -292,7 +292,7 @@ def test_switch_between_wip_and_version(user_alice_logged_in, handle_usage_stati
     #
     surface = SurfaceFactory(creator=user_alice, name="Alice's Surface")
     topo = Topography2DFactory(surface=surface)
-    publication = surface.publish('cc0-1.0', 'Alice')
+    publication = Publication.publish(surface, 'cc0-1.0', 'Alice')
 
     #
     # When browsing to the surface, the current surface is shown as "Work in progress"
@@ -362,7 +362,7 @@ def test_how_to_cite(user_alice_logged_in, handle_usage_statistics):
 
     surface = SurfaceFactory(creator=user_alice, name=surface_name)
     topo = Topography2DFactory(surface=surface, name=topo_name, description=topo_description)
-    publication = surface.publish('cc0-1.0', 'Famous Scientist')
+    publication = Publication.publish(surface, 'cc0-1.0', 'Famous Scientist')
 
     # Alice filters for published surfaces - enters
     # "Select" tab and chooses "Only published surfaces"
