+++ topobank/manager/management/commands/fix_permissions.py
@@ -5,6 +5,7 @@ from django.core.management.base import BaseCommand
 from guardian.shortcuts import assign_perm
 
 from topobank.manager.models import Surface
+from topobank.publication.utils import set_publication_permissions
 
 _log = logging.getLogger(__name__)
 
@@ -44,7 +45,7 @@ class Command(BaseCommand):
                 if any(creator.has_perm(p, surface) for p in ALL_PERMISSIONS):
                     if not options['dry_run']:
                         # there should be no individual rights for published surfaces
-                        surface.set_publication_permissions()
+                        set_publication_permissions(surface)
                     fixed_published_surfaces.add(surface)
             else:
                 for perm in ALL_PERMISSIONS:
